requirements.txt
flask
tensorflow
tensorflowjs
numpy
pillow